requirements.txt
discogs-client==2.3.0
python-dotenv==1.0.0
urllib3==2.0.7 